use the current libproj/PROJ API
The traditional PROJ4 API has now been obsolete for a number of years;
port to the current PROJ API (using PROJ 7.2 while doing so).

I get the impression that the new API will be more volatile: the new
maintainer seems happy to break things, in the way that is fascionable
nowadays.

The code still uses at least one semi-obsolete feature, namely
referring to RT90 and SWEREF99 under the names epsg:3021 and epsg:3006.

    It's not clear from the documentation if/when that feature will
    disappear completely, or what the alternative is. The project
    seems to want users to depend on Internet connectivity to their
    own database. Which of course is completely unacceptable.

The porting has unfortunately come with decreased accuracy. A trans-
formation error used to be measured in millimeters; now it's a
decimeter or so. It's unclear to me if it's my fault. The new PROJ
maintainers would probably say it is, for not calling some under-
documented function, passing some exotic parameter, or taken a
university course in cartography. Anyway, for typical use of rt90 --
with coordinates found on a map or estimated by a GPS -- this has no
real significance, so I let it slide.