rework command-line parsing
What I've done in recent projects: collect the arguments in a
locally defined, anonymous struct, pass it around to functions which
take a class Arg as template argument.  Avoids dragging around N
separate arguments which modify the processing, but doesn't make
this struct a too prominent type.

It's a good compromise between control and informality.